link cell objects to clue numbers
When a user click on a cell of a grid we want to highlight the down and
across clue associated with that cell. To do this quickly, we process
the board into a "playable" board. This takes an array of arrays of string
and maps them to an array of array of objects. these "cell" objects have
the shape {answer, guess, number, [acrossClueNumber, downClueNumber]}

Saving this information allows us to highligh the clues instantly
without having to do any processing to figure out which clues are
supposed to be highlighted.